requirements.txt
streamlit
requests
python-dotenv
textblob
googletrans==4.0.0-rc1